requirements.txt
django==1.7
requests==1.2.3
django-tastypie==0.11.0
gunicorn==19.0.0
psycopg2==2.5.3
static3==0.5.1
wsgiref==0.1.2